Move debug GUI out of legacy
