+++ exp_scenario/module/control.lua
@@ -81,3 +81,5 @@ add(require("modules/exp_scenario/gui/rocket_info"))
 add(require("modules/exp_scenario/gui/science_production"))
 add(require("modules/exp_scenario/gui/surveillance"))
 add(require("modules/exp_scenario/gui/task_list"))
+add(require("modules/exp_scenario/gui/debug/shim"))
+add(require("modules/exp_scenario/gui/debug/event_view"))
