+++ exp_scenario/module/gui/debug/main_view.lua
@@ -1,16 +1,16 @@
-local Gui = require("modules.exp_legacy.utils.gui") --- @dep utils.gui
+local Gui = require("modules/exp_scenario/gui/debug/shim")
 local Color = require("modules/exp_util/include/color")
 
 local Public = {}
 
 local pages = {
-    require("modules.exp_legacy.modules.gui.debug.redmew_global_view"),
-    require("modules.exp_legacy.modules.gui.debug.expcore_datastore_view"),
-    require("modules.exp_legacy.modules.gui.debug.expcore_gui_view"),
-    require("modules.exp_legacy.modules.gui.debug.global_view"),
-    require("modules.exp_legacy.modules.gui.debug.package_view"),
-    require("modules.exp_legacy.modules.gui.debug._g_view"),
-    require("modules.exp_legacy.modules.gui.debug.event_view"),
+    require("modules/exp_scenario/gui/debug/redmew_global_view"),
+    require("modules/exp_scenario/gui/debug/expcore_datastore_view"),
+    require("modules/exp_scenario/gui/debug/expcore_gui_view"),
+    require("modules/exp_scenario/gui/debug/global_view"),
+    require("modules/exp_scenario/gui/debug/package_view"),
+    require("modules/exp_scenario/gui/debug/_g_view"),
+    require("modules/exp_scenario/gui/debug/event_view"),
 }
 
 local main_frame_name = Gui.uid_name()
